requirements.txt
graphene-django
djangorestframework